+++ src/proto/message.pro
@@ -73,5 +73,4 @@ int do_dialog(int type, char_u *title, char_u *message, char_u *buttons, int dfl
 int vim_dialog_yesno(int type, char_u *title, char_u *message, int dflt);
 int vim_dialog_yesnocancel(int type, char_u *title, char_u *message, int dflt);
 int vim_dialog_yesnoallcancel(int type, char_u *title, char_u *message, int dflt);
-char_u *do_browse(int flags, char_u *title, char_u *dflt, char_u *ext, char_u *initdir, char_u *filter, buf_T *buf);
 /* vim: set ft=c : */
